+++ projects/ngx-amap/src/types/class/district-search/district-search-result.ts
@@ -4,4 +4,3 @@ export interface DistrictSearchResult {
   info: string;
   districtList: District[];
 }
-