+++ projects/ngx-amap/src/types/class/transfer/segment.ts
@@ -1,0 +1,7 @@
+export interface Segment {
+  instruction: string;
+  transit_mode: string;
+  time: number;
+  transit: any; // TODO: TransitDetail | WalkDetails
+  distance: number;
+}