+++ projects/ngx-amap/src/types/interface/walking/walking-options.interface.ts
@@ -1,0 +1,10 @@
+import { Map } from '../../class/amap.map';
+
+export interface WalkingOptions {
+  map?: Map;
+  panel?: any;
+  hideMarkers?: boolean;
+  isOutline?: boolean;
+  outlineColor?: string;
+  autoFitView?: boolean;
+}