+++ src/app/_nav.ts
@@ -322,6 +322,38 @@ export const navigation = [
     name: '服务示例'
   },
   {
+    name: '路径规划服务',
+    url: '/',
+    icon: 'fa fa-cogs',
+    children: [
+      {
+        name: '驾车路线',
+        url: '/AmapDrivingService/simple',
+        icon: 'fa fa-html5'
+      },
+      {
+        name: '货车路线',
+        url: '/AmapTruckDrivingService/simple',
+        icon: 'fa fa-html5'
+      },
+      {
+        name: '公交换乘路线',
+        url: '/AmapTransferService/simple',
+        icon: 'fa fa-html5'
+      },
+      {
+        name: '步行路线',
+        url: '/AmapWalkingService/simple',
+        icon: 'fa fa-html5'
+      },
+      {
+        name: '骑行路线',
+        url: '/AmapRidingService/simple',
+        icon: 'fa fa-html5'
+      },
+    ]
+  },
+  {
     name: 'AmapGeocoderService',
     url: '/AmapGeocoderService',
     icon: 'fa fa-cogs',