+++ src/app/views/amap-truck-driving-demo/amap-truck-driving-demo.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { SharedModule } from '../shared/shared.module';
+import { AmapTruckDrivingDemoRoutingModule } from './amap-truck-driving-demo-routing.module';
+import { SimpleComponent } from './simple/simple.component';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    SharedModule,
+    AmapTruckDrivingDemoRoutingModule
+  ],
+  declarations: [
+    SimpleComponent
+  ]
+})
+export class AmapTruckDrivingDemoModule { }